requirements.txt
# requirements.txt

streamlit
crewai
crewai_tools
langchain-google-genai
python-dotenv
pymupdf